+++ examples/component-browser/bottles-library.js
@@ -1,4 +1,3 @@
-/* jslint node: true */
 "use strict";
 
 var Yadda = require('yadda');