+++ examples/context/bin/example.js
@@ -1,4 +1,3 @@
-/* jslint node: true */
 "use strict";
 
 var spawn = require('child_process').spawn;