+++ examples/mocha-express/app.js
@@ -1,4 +1,3 @@
-/* jslint node: true */
 "use strict";
 
 var express = require('express');