+++ examples/mocha-express/index.js
@@ -1,4 +1,3 @@
-/* jslint node: true */
 "use strict";
 
 require('./app').start('localhost', 3000);